pyproject.toml
[tool.poetry]
name = "piazza-attendant"
version = "0.1.0"
description = ""
authors = ["aura-ix <173636334+aura-ix@users.noreply.github.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
piazza-api = "^0.14.3"
python-dateutil = "^2.9.0.post0"
requests = "^2.32.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
